fix: tests, dyn allocator
